Lift matched-filter stacking and orbit metric into p9-core
